Chakra UI Setup

+++ backend/controllers/product.controller.js
@@ -71,6 +71,12 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({
+      success: "false",
+      message: "Invalid product ID",
+    });
+  }
 
   try {
     await Product.findByIdAndDelete(id);
@@ -79,9 +85,9 @@ export const deleteProduct = async (req, res) => {
       message: "Product deleted successfully",
     });
   } catch (error) {
-    res.status(404).json({
+    res.status(500).json({
       success: "false",
-      message: "Product not found",
+      message: "Server Error",
     });
   }
 };
